Undo system
What we do is we have the undo system implemented as a stack of show snap shots. When we do an undo, we restore the show to a previous state.

With CalChart, not every Store commit is a change we want to undo, so we have a filter where we only save certain commits.

Adding bottom row buttons

using hotkeys.

adding File/Edit menu.

Fixing up linting and tests.